기존의 profile을 compose 부품으로 변환
